Fixes ALF-9220: WQS: Calling "getSize()" on an Asset throws a ClassCastException
git-svn-id: http://svn.alfresco.com/repos/alfresco-open-mirror/alfresco/HEAD@28573 7033f55c-6c29-4ad8-afba-765cb616e738

+++ root/modules/wcmquickstart/wcmquickstartclientapi/source/java/org/alfresco/wcm/client/impl/ResourceBaseImpl.java
@@ -18,6 +18,7 @@
 package org.alfresco.wcm.client.impl;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.Map;
 import java.util.TreeMap;
 
@@ -205,7 +206,7 @@ public abstract class ResourceBaseImpl implements Resource
             ContentInfo contentInfo = (ContentInfo) properties.get("cm:content");
             if (contentInfo != null)
             {
-                properties.put(PropertyIds.CONTENT_STREAM_LENGTH, contentInfo.getSize());
+                properties.put(PropertyIds.CONTENT_STREAM_LENGTH, BigInteger.valueOf(contentInfo.getSize()));
                 properties.put(PropertyIds.CONTENT_STREAM_MIME_TYPE, contentInfo.getMimeType());
             }
         }
